+++ tests/test_iscc_sdk.py
@@ -2,4 +2,4 @@ from iscc_sdk import __version__
 
 
 def test_version():
-    assert __version__ == "0.5.8"
+    assert __version__ == "0.5.9"